+++ experimental/torchprep/tests/config/bert.yaml
@@ -6,14 +6,14 @@
 
 # BERT model
 input1:
-  dtype: "int32" # Extend this to a list later, will need to print all permutations
+  dtype: "int32"
   device: "cpu"
   shape: [1] # the first element is the batch size
   high : 10 # max value
   mode : "latency"
 
 input2:
-  dtype: "float32" # Extend this to a list later, will need to print all permutations
+  dtype: "float32"
   device: "cpu"
   shape: [1] # the first element is the batch size
   high : 10 # max value
